Colocated PS

+++ daiet/src/params.cpp
@@ -47,7 +47,7 @@ namespace daiet {
                 ("daiet.max_num_pending_messages", po::value<uint32_t>(&(daiet_par.getMaxNumPendingMessages()))->default_value(256), "Max number of pending, unaggregated messages")
                 ("daiet.num_updates", po::value<uint32_t>(&num_updates)->default_value(32), "Number of updates per packet")
 #ifdef COLOCATED
-                ("daiet.num_workers", po::value<uint32_t>(&(daiet_par.getNumWorkers()))->default_value(0), "Number of workers")
+                ("daiet.num_workers", po::value<uint16_t>(&(daiet_par.getNumWorkers()))->default_value(0), "Number of workers")
 #endif
 #ifdef TIMERS
                 ("daiet.timeout", po::value<double>(&(daiet_par.getTimeout()))->default_value(1), "Timeout in millisecond")
@@ -170,7 +170,7 @@ namespace daiet {
     }
 
 #ifdef COLOCATED
-    uint32_t& daiet_params::getNumWorkers() {
+    uint16_t& daiet_params::getNumWorkers() {
         return num_workers;
     }
 #endif
